feat : update all fiture

+++ app/Http/Controllers/api/CustomerController.php
@@ -2,6 +2,7 @@
 
     namespace App\Http\Controllers\Api;
     
+    use App\Models\Customer;
     use App\DTOs\CustomerDTO;
     use Illuminate\Http\Request;
     use App\Services\CustomerService;
@@ -9,10 +10,12 @@
     use App\Http\Controllers\Controller;
     use App\Http\Requests\CustomerRequest;
     use App\Http\Resources\CustomerResource;
+    use Illuminate\Support\Facades\Gate;
     
     class CustomerController extends Controller
     {
         use ApiResponse;
+        public $model = Customer::class;
     
         public function __construct(
             protected CustomerService $service
@@ -20,9 +23,12 @@
     
         public function index(Request $request)
         {
+            Gate::authorize('viewAny', $this->model);
+
             return $this->handleException(function () use ($request) {
     
-                $list = $this->service->list($request->all());
+                $where = ['owner_id' => $this->ownerid($request)];
+                $list = $this->service->list($request->all(), $where);
             
                 return $this->success_list(
                     CustomerResource::collection($list['data']),
@@ -39,13 +45,16 @@
     
         public function store(CustomerRequest $request)
         {
+            Gate::authorize('create', $this->model);
+
             return $this->handleException(function () use ($request) {
-                $data = $this->service->create(
-                    UserDTO::fromArray($request->validated())
-                );
+
+                $dto = CustomerDTO::fromArray($request->validated());
+                $dto->owner_id = $this->ownerid($request);
+                $data = $this->service->create($dto);
     
                 return $this->success(
-                    new UserResource($data),
+                    new CustomerResource($data),
                     'Customer created',
                     201
                 );
@@ -54,12 +63,14 @@
     
         public function show(int $id)
         {
+            Gate::authorize('view', $this->model);
+
             return $this->handleException(function () use ($id) {
     
                 $data = $this->service->find($id);
     
                 return $this->success(
-                    new UserResource($data),
+                    new CustomerResource($data),
                     'Customer detail retrieved'
                 );
             });
@@ -67,11 +78,15 @@
     
         public function update(CustomerRequest $request, int $id)
         {
-            return $this->handleException(function () use ($id) {
+            Gate::authorize('update', $this->model);
+
+            return $this->handleException(function () use ($id, $request) {
     
+                $dto = CustomerDTO::fromArray($request->validated());
+                $dto->owner_id = $this->ownerid($request);
                 $data = $this->service->update(
                     $id,
-                    CustomerDTO::fromArray($request->validated())
+                    $dto
                 );
     
                 return $this->success(
@@ -83,6 +98,8 @@
     
         public function destroy(int $id)
         {
+            Gate::authorize('delete', $this->model);
+
             return $this->handleException(function () use ($id) {
     
                 $this->service->delete($id);
